+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version 1.1-SNAPSHOT
 ADDED:
 
 - new callback methods in action classes
+- new factory class for create all components for the help window
+- new extension methods in class MenuExtensions
 
 Version 1
 -------------